Removed feature that deleted triangles just after the reload of the new ones.
Now old triangles are still drawed in the screen, but when a new triangle is drawn over them they are stored in a list of the model.

A new button was added to the interface with the name of "optimize GPU memory" that deletes the triangles stored in the list of the model.