Added shortcut feature to open a netcdf file.
To make this possible it was necesary to refactor the code, changin the logic of the load of files from the gui to the program, and thus, was necesary to implement a new way of showing modals on the program

The GUI manager now have a new method to create modals on the program with custom text.

Due to the refactor of the logic, it was necesary to add routines of try-except in differents levels of the program (engine, gui, controller).

Added the tools and the shortcuts to the readme.