Added shortcut to the change of cpt files.
TO make this, the controller was modified, adding ne calls to the engine to change the cpt file and with it the models.

The engine calls the program, and is the program the one who does the logic of the operation.

The program opens a dialog to select a file (from here exceptions occurs), and tells the engine to update the models once the cpt file was chasnged.

there is a lot of exception when changing the files, they are caught in both places where the call to change the cpt can happen, in the gui manager, and in the controller. In both cases a modals pop up and show an error message.