Added functionality to draw the points of the polygons using the new Points class, also updated the engine and the polygon class to do this task.
The engine now calls for the active_polygon_line_width variable to the settings.

Created two new shaders that are used by the points models.

Polygon now only stores the color to use for the lines.
Polygon now doesnt change the uniform variable on the draw method.
Polygon now add a point to the instance of Points model when added a new point to the polygon.py
Polygon now call the draw method of the Points instance instead of drawing itself.
Polygon now change the color of the Point instance to change the color of the dots of the polygons.