Added ctrl+z functionality to the program related to the polygons.
Now the polygons delete the last added point when ctrl z is pressed.

The polygons keep the state from when they were created, not impacting in the other proccesses of the program.

Modified readme and class diagram.

The polygon models also updated the points model that the first one use for drawing the points.