Updated documentation of package and class.
Added functionality to check in real time if name of a parameter is repaeated or not in the creation of a new paramter of a polygon, showing a message to the user if the name is repeated.

Button add parameter does nothing if the name of the parameter is repeated.