Read netcdf file will always give ascending arrays.
Modified the input module to make that the arrays read from netcdf files are always returned in ascending order.

Modified tests that tested the read and storing of the arrays to make sense of the new way of storing the arrays.

Modified method that made a check if the arrays where ascending or descending.