OOP Fonts (#20)
* Added new Font enum class.

* Updated usage of the Font class.

Changed all the methods that were defined to change fonts to one method set_font on the guimanager.

Updated usage of the deleted methods.

Removed parameters from the constructor of the class GUIMAnager that were associated with the fonts. Also removed methods and parameters from the other classes associated with the fonts.

* Renamed TOOL_TILE to TOOL_TITLE.

Updated usages that were bugged.