Added new transformation.
Added new map transformation to the program to eliminate the values in one map (replace them with nan)  that are defined in another map.

Created a new modal class for the configuration and application of the new transformation.

Modified mainmenubar to add a new item to open the created modal.

Added tests for the new map transformation.